refactor: extract RenderCore base with determinism seams
Move all non-drawing state and behavior (interaction, frame scheduling,
hit-testing, signal spawning, options, projCache) from BrainRenderer into
an abstract RenderCore base class. BrainRenderer now extends RenderCore and
implements drawScene() for Canvas2D rendering. Adds setDeterministic/setView
seams for the upcoming A/B WebGL harness.
